Fix/crusial

+++ mlcomp/server/front/src/app/model/model.component.ts
@@ -11,6 +11,8 @@ import {
 import {Helpers} from "../helpers";
 import {DomSanitizer} from "@angular/platform-browser";
 import {ModelStartDialogComponent} from "./model-start-dialog.component";
+import {ModelAddDialogComponent} from "./model-add-dialog.component";
+import {DialogComponent} from "../dialog/dialog.component";
 
 @Component({
     selector: 'app-model',
@@ -23,8 +25,6 @@ export class ModelComponent extends Paginator<Model> {
         'dag',
         'id',
         'name',
-        'slot',
-        'interface',
         'created',
         'score_local',
         'score_public',
@@ -47,6 +47,8 @@ export class ModelComponent extends Paginator<Model> {
                 iconRegistry: MatIconRegistry,
                 sanitizer: DomSanitizer,
                 public start_dialog: MatDialog,
+                public model_add_dialog: MatDialog,
+                public dialog: MatDialog
     ) {
         super(service, location);
 
@@ -67,7 +69,7 @@ export class ModelComponent extends Paginator<Model> {
     protected _ngOnInit() {
         let self = this;
         this.data_updated.subscribe(res => {
-            if(!res || !res.projects){
+            if (!res || !res.projects) {
                 return;
             }
             self.projects = res.projects;
@@ -104,33 +106,46 @@ export class ModelComponent extends Paginator<Model> {
 
     remove(element: Model) {
         let self = this;
-        this.service.remove(element.id).subscribe(res => {
-            self.change.emit();
+        const dialogRef = this.dialog.open(DialogComponent, {
+            width: '550px', height: '200px',
+            data: {
+                'message': 'The all content will be deleted. ' +
+                    'Do you want to continue?'
+            }
         });
+
+        dialogRef.afterClosed().subscribe(result => {
+            if (result) {
+                this.service.remove(element.id).subscribe(res => {
+                    self.change.emit();
+                });
+            }
+        })
     }
 
 
     start(element: Model) {
-        let dag;
-        for (let d of element.dags) {
-            if (element.dag == d.id) {
-                dag = d;
-                break
-            }
-        }
         let config = {
-            width: '500px', height: '400px',
+            width: '1050px', height: '800px',
             data: {
-                'dags': element.dags,
-                'interface': element.interface,
-                'slot': element.slot,
-                'dag': dag,
-                'interface_params': element.interface_params,
                 'model_id': element.id
             }
         };
         let dialog = this.start_dialog.open(ModelStartDialogComponent, config);
-        dialog.afterClosed().subscribe(res=>this.change.emit());
+        dialog.afterClosed().subscribe(res => this.change.emit());
+    }
+
+    add() {
+        let dialog = this.model_add_dialog.open(ModelAddDialogComponent,
+            {
+                width: '500px', height: '400px',
+                data: {
+                    'projects': this.projects.filter(x => x.name != 'None'),
+                    'equations': '',
+                    'name': ''
+                }
+            });
+        dialog.afterClosed().subscribe(res => this.change.emit());
     }
 }
